requirements.txt
absl-py==1.3.0
apache-beam==2.43.0
astunparse==1.6.3
attrs==22.2.0
avro-python3==1.10.2
cachetools==5.2.1
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.6
contextlib2==21.6.0
contourpy==1.0.6
crcmod==1.7
cycler==0.11.0
Cython==0.29.33
dill==0.3.1.1
dm-tree==0.1.8
docopt==0.6.2
etils==0.9.0
fastavro==1.7.0
fasteners==0.18
flatbuffers==23.1.4
fonttools==4.38.0
gast==0.4.0
gin-config==0.5.0
google-api-core==2.11.0
google-api-python-client==2.71.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.57.1
grpcio==1.51.1
h5py==3.7.0
hdfs==2.7.0
httplib2==0.20.4
idna==3.4
immutabledict==2.2.3
importlib-resources==5.10.2
joblib==1.2.0
kaggle==1.5.12
keras==2.11.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==15.0.6.1
lvis==0.5.3
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
neural-structured-learning==1.4.0
numpy==1.24.1
oauth2client==4.1.3
oauthlib==3.2.2
object-detection @ file:///C:/Users/ward_/Downloads/TensorFlow/models/research
objsize==0.5.2
opencv-python==4.7.0.68
opencv-python-headless==4.7.0.68
opt-einsum==3.3.0
orjson==3.8.4
packaging==23.0
pandas==1.5.2
Pillow==9.4.0
portalocker==2.6.0
promise==2.3
proto-plus==1.22.1
protobuf==3.19.6
psutil==5.9.4
psycopg2-binary==2.9.5
py-cpuinfo==9.0.0
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pydot==1.4.2
pymongo==3.13.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-slugify==7.0.0
pytz==2022.7
pywin32==305
PyYAML==5.4.1
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
sacrebleu==2.2.0
scikit-learn==1.2.0
scipy==1.10.0
sentencepiece==0.1.97
seqeval==1.2.2
six==1.16.0
tabulate==0.9.0
tensorboard==2.11.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-addons==0.19.0
tensorflow-datasets==4.8.1
tensorflow-estimator==2.11.0
tensorflow-hub==0.12.0
tensorflow-intel==2.11.0
tensorflow-io==0.29.0
tensorflow-io-gcs-filesystem==0.29.0
tensorflow-metadata==1.12.0
tensorflow-model-optimization==0.7.3
tensorflow-text==2.10.0
termcolor==2.2.0
text-unidecode==1.3
tf-models-official==2.10.1
tf-slim==1.1.0
threadpoolctl==3.1.0
toml==0.10.2
tqdm==4.64.1
typeguard==2.13.3
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
Werkzeug==2.2.2
wincertstore==0.2
wrapt==1.14.1
zipp==3.11.0
zstandard==0.19.0
